// src/app/core/analytics/resources/events-interfaces/edit-item-re.interface.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * User edits an uploaded product (in real estate)
 */
export interface EditItemRE {
  /**
   * The unique identifier for an item
   */
  itemId: string;
  /**
   * Identifier of the category that the item belongs to
   */
  categoryId: number;
  /**
   * Item price
   */
  salePrice: number;
  /**
   * Item title
   */
  title?: string;
  /**
   * Real State Operation
   */
  operation?: string;
  /**
   * Real State Property Type
   */
  type?: string;
  /**
   * Real State condition
   */
  condition?: string;
  /**
   * Real State surface area
   */
  surface?: number;
  /**
   * Real State number of rooms
   */
  rooms?: number;
  /**
   * If the item edited is from a professional user
   */
  isPro: boolean;
  /**
   * Identifier of the screen that the item was edited from
   */
  screenId: 162;
}